warehouse/app/requirements.txt
# Application requirements
flask==2.2.5
flask-restx==1.1.0
flask-sqlalchemy==3.0.5
pytest-flask==1.2.0
freezegun==1.2.2
delorean==1.0.0
faker==18.13.0
PyJWT==2.8.0
cryptography==41.0.2
parse==1.19.1
Werkzeug==2.2.3
flake8==3.7.7